--- clients/spark/project/types.scala
@@ -1,9 +1,0 @@
-package build
-
-class BuildType(
-    val suffix: String,
-    val sparkVersion: String,
-    val scalapbVersion: String,
-    val hadoopFlavour: String, // If set, a directory of additional sources to compile
-    val gcpConnectorVersion: String
-)
