+++ design/accepted/export-functionality.md
@@ -27,7 +27,7 @@ Usage example:
 docker run lakefs-export --conf spark.hadoop.lakefs.api.url=https://<LAKEFS_ENDPOINT>/api/v1 \
 --conf spark.hadoop.lakefs.api.access_key=<LAKEFS_ACCESS_KEY_ID> \
 --conf spark.hadoop.lakefs.api.secret_key=<LAKEFS_SECRET_ACCESS_KEY> \
---packages io.lakefs:lakefs-spark-client-301_2.12:0.1.0 \
+--packages io.lakefs:lakefs-spark-client_2.12:0.1.0 \
 --class io.treeverse.clients.Main export-app example-repo s3://example-bucket/prefix \
 --branch=example-branch 
 ```
